fix: ProductController, BrandController 사용자 권한에 따른 클래스 분리
